backend/requirements.txt
# Minimal requirements for Render deployment (no compilation)
fastapi==0.128.3
uvicorn[standard]==0.30.0
python-dotenv==1.2.2
python-multipart==0.0.22
numpy==1.26.4
pandas==3.0.0
scikit-learn==1.5.2
scipy==1.17.1
google-genai==1.0.0
shap==0.46.0